# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CineClub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{username}:
    get:
      operationId: getUserByUsername
      tags: [users]
      summary: Get user by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found

  # ── Movies (TMDB proxy) ────────────────────────────────────────────────────
...
  /watchlists:
    get:
      operationId: getWatchlists
      tags: [watchlists]
      summary: List current user's watchlists
      responses:
        "200":
          description: Watchlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Watchlist"
    post:
      operationId: createWatchlist
      tags: [watchlists]
      summary: Create a watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
  /watchlists/{id}:
    patch:
      operationId: updateWatchlist
      tags: [watchlists]
      summary: Update watchlist name
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
    delete:
      operationId: deleteWatchlist
      tags: [watchlists]
      summary: Delete a watchlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /watchlists/{id}/movies:
    get:
      operationId: getWatchlistMovies
      tags: [watchlists]
      summary: List movies in a watchlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistMovie"
    post:
      operationId: addMovieToWatchlist
      tags: [watchlists]
      summary: Add a movie to a watchlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistMovieInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistMovie"
  /watchlists/{id}/movies/{tmdbId}:
    delete:
      operationId: removeMovieFromWatchlist
      tags: [watchlists]
      summary: Remove a movie from a watchlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: tmdbId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Reviews ────────────────────────────────────────────────────────────────
  /reviews:
    post:
      operationId: upsertReview
      tags: [reviews]
      summary: Create or update a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "200":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/movie/{tmdbId}:
    get:
      operationId: getMovieReviews
      tags: [reviews]
      summary: Get all reviews for a movie
      parameters:
        - name: tmdbId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewWithUser"
  /reviews/user/{userId}:
    get:
      operationId: getUserReviews
      tags: [reviews]
      summary: Get reviews by a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
  /reviews/{id}:
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Friends ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, username, displayName, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        watchedCount:
          type: integer
        reviewCount:
          type: integer
        friendCount:
          type: integer
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
...
    Watchlist:
      type: object
      required: [id, userId, name, isDefault, movieCount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        isDefault:
          type: boolean
        movieCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    WatchlistInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
    WatchlistUpdate:
      type: object
      required: [name]
      properties:
        name:
          type: string
    WatchlistMovie:
      type: object
      required: [id, watchlistId, tmdbId, title, posterPath, addedAt]
      properties:
        id:
          type: integer
        watchlistId:
          type: integer
        tmdbId:
          type: integer
        title:
          type: string
        posterPath:
          type: ["string", "null"]
        releaseYear:
          type: ["integer", "null"]
        voteAverage:
          type: number
        addedAt:
          type: string
          format: date-time
    WatchlistMovieInput:
      type: object
      required: [tmdbId, title]
      properties:
        tmdbId:
          type: integer
        title:
          type: string
        posterPath:
          type: ["string", "null"]
        releaseYear:
          type: ["integer", "null"]
        voteAverage:
          type: number
    Review:
      type: object
      required: [id, userId, tmdbId, rating, content, isSpoiler, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        tmdbId:
          type: integer
        movieTitle:
          type: ["string", "null"]
        moviePosterPath:
          type: ["string", "null"]
        rating:
          type: number
        content:
          type: string
        isSpoiler:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ReviewWithUser:
      type: object
      required: [id, userId, tmdbId, rating, content, isSpoiler, createdAt, user]
      properties:
        id:
          type: integer
        userId:
          type: string
        tmdbId:
          type: integer
        rating:
          type: number
        content:
          type: string
        isSpoiler:
          type: boolean
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/UserProfile"
    ReviewInput:
      type: object
      required: [tmdbId, rating, content, isSpoiler]
      properties:
        tmdbId:
          type: integer
        movieTitle:
          type: string
        moviePosterPath:
          type: ["string", "null"]
        rating:
          type: number
        content:
          type: string
        isSpoiler:
          type: boolean
...