# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CineClub API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  userId: string;
  tmdbId: number;
  /** @nullable */
  movieTitle?: string | null;
  /** @nullable */
  moviePosterPath?: string | null;
  rating: number;
  content: string;
  isSpoiler: boolean;
  createdAt: Date;
}