# lib/api-zod/src/generated/types/reviewWithUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CineClub API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from './userProfile';

export interface ReviewWithUser {
  id: number;
  userId: string;
  tmdbId: number;
  rating: number;
  content: string;
  isSpoiler: boolean;
  createdAt: Date;
  user: UserProfile;
}