# lib/api-zod/src/generated/types/watchlist.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CineClub API
 * OpenAPI spec version: 0.1.0
 */

export interface Watchlist {
  id: number;
  userId: string;
  name: string;
  isDefault: boolean;
  movieCount: number;
  createdAt: Date;
}