# lib/api-zod/src/generated/types/watchlistMovie.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CineClub API
 * OpenAPI spec version: 0.1.0
 */

export interface WatchlistMovie {
  id: number;
  watchlistId: number;
  tmdbId: number;
  title: string;
  /** @nullable */
  posterPath: string | null;
  /** @nullable */
  releaseYear?: number | null;
  voteAverage?: number;
  addedAt: Date;
}